Update public keys table
